requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
boto3==1.34.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests==2.31.0
gunicorn==21.2.0
blinker==1.7.0
